refactor(ui): move SectionHeader to shared UI components
