+++ app/[language]/(guest)/(default-layout)/_home/Slogan/SloganSection.tsx
@@ -1,5 +1,5 @@
 import { cn } from '@/utils/cn';
-import { SectionHeader } from '@/app/[language]/(guest)/(default-layout)/SectionHeader';
+import { SectionHeader } from '@/shared/ui/section-header';
 import Image from 'next/image';
 
 interface SloganSectionProps {
